News & Events, Regulations and Contact Us pages done
The about us pages are done, the news and events page is done, even the regulations pages, the controller for it was also done; now only the about us in the menu has to be drop down so we can link the pages.
Contact Us page had to be done again since the last time wasn't saved. It's CSS has to be edit for minor changes.